requirements.txt
fastapi[standard]==0.116.1
python-telegram-bot==22.3
python-decouple==3.8
sqlmodel==0.0.24
async-lru==2.0.5